Moved useScope to own module to eliminate circular dependency
