+++ lib/main.ts
@@ -2,7 +2,7 @@ import { createContext } from "./context.ts";
 import { type Operation } from "./types.ts";
 import { callcc } from "./callcc.ts";
 import { run } from "./run.ts";
-import { useScope } from "./scope.ts";
+import { useScope } from "./use-scope.ts";
 import process from "node:process";
 
 /**
